+++ EveLib.EveXml/Models/Misc/CallList.cs
@@ -87,7 +87,7 @@ namespace eZet.EveLib.EveXmlModule.Models.Misc {
             /// </summary>
             /// <value>The access mask.</value>
             [XmlAttribute("accessMask")]
-            public int AccessMask { get; set; }
+            public long AccessMask { get; set; }
 
             /// <summary>
             ///     Gets or sets the character.
